requirements.txt
streamlit
pandas
psycopg2-binary
openpyxl
reportlab
streamlit-cookies-manager
cryptography
streamlit-option-menu